+++ tests/test_providers.py
@@ -1,7 +1,9 @@
 from typing import Any, List, Set, Callable, Sequence
+from unittest import mock
 
 import attr
 import json
+import pytest
 from pytest_twisted import inlineCallbacks
 from scrapy_poet import ResponseDataProvider
 from twisted.python.failure import Failure
@@ -11,9 +13,14 @@ from scrapy import Request, Spider
 from scrapy.crawler import Crawler
 from scrapy.settings import Settings
 from scrapy.utils.test import get_crawler
-from scrapy_poet.page_input_providers import CacheDataProviderMixin, PageObjectInputProvider
-from tests.utils import crawl_single_item, HtmlResource
-from web_poet import ResponseData
+from scrapy_poet.page_input_providers import (
+    CacheDataProviderMixin,
+    PageObjectInputProvider,
+    HttpClientProvider,
+    MetaProvider,
+)
+from tests.utils import crawl_single_item, HtmlResource, AsyncMock
+from web_poet import ResponseData, HttpClient
 
 
 class ProductHtml(HtmlResource):
@@ -28,6 +35,7 @@ class ProductHtml(HtmlResource):
         <p class="description">The best chocolate ever</p>
     </html>
     """
+
 
 class NonProductHtml(HtmlResource):
     html = """
@@ -61,7 +69,9 @@ class PriceHtmlDataProvider(PageObjectInputProvider, CacheDataProviderMixin):
         assert isinstance(crawler, Crawler)
         super().__init__(crawler)
 
-    def __call__(self, to_provide, response: scrapy.http.Response, spider: scrapy.Spider):
+    def __call__(
+        self, to_provide, response: scrapy.http.Response, spider: scrapy.Spider
+    ):
         assert isinstance(spider, scrapy.Spider)
         ret: List[Any] = []
         if Price in to_provide:
@@ -128,7 +138,14 @@ class PriceFirstMultiProviderSpider(scrapy.Spider):
     def errback(self, failure: Failure):
         yield {"exception": failure.value}
 
-    def parse(self, response, price: Price, name: Name, html: Html, response_data: ResponseData):
+    def parse(
+        self,
+        response,
+        price: Price,
+        name: Name,
+        html: Html,
+        response_data: ResponseData,
+    ):
         yield {
             Price: price,
             Name: name,
@@ -152,8 +169,9 @@ class NameFirstMultiProviderSpider(PriceFirstMultiProviderSpider):
 def test_name_first_spider(settings, tmp_path):
     cache = tmp_path / "cache.sqlite3"
     settings.set("SCRAPY_POET_CACHE", str(cache))
-    item, _, _ = yield crawl_single_item(NameFirstMultiProviderSpider, ProductHtml,
-                                         settings)
+    item, _, _ = yield crawl_single_item(
+        NameFirstMultiProviderSpider, ProductHtml, settings
+    )
     assert cache.exists()
     assert item == {
         Price: Price("22€"),
@@ -164,8 +182,9 @@ def test_name_first_spider(settings, tmp_path):
 
     # Let's see that the cache is working. We use a different and wrong resource,
     # but it should be ignored by the cached version used
-    item, _, _ = yield crawl_single_item(NameFirstMultiProviderSpider, NonProductHtml,
-                                         settings)
+    item, _, _ = yield crawl_single_item(
+        NameFirstMultiProviderSpider, NonProductHtml, settings
+    )
     assert item == {
         Price: Price("22€"),
         Name: Name("Chocolate"),
@@ -174,11 +193,11 @@ def test_name_first_spider(settings, tmp_path):
     }
 
 
-
 @inlineCallbacks
 def test_price_first_spider(settings):
-    item, _, _ = yield crawl_single_item(PriceFirstMultiProviderSpider, ProductHtml,
-                                         settings)
+    item, _, _ = yield crawl_single_item(
+        PriceFirstMultiProviderSpider, ProductHtml, settings
+    )
     assert item == {
         Price: Price("22€"),
         Name: Name("Chocolate"),
@@ -195,3 +214,33 @@ def test_response_data_provider_fingerprint(settings):
     # The fingerprint should be readable since it's JSON-encoded.
     fp = rdp.fingerprint(scrapy.http.Response, request)
     assert json.loads(fp)
+
+
+@pytest.mark.asyncio
+async def test_http_client_provider(settings):
+    crawler = get_crawler(Spider, settings)
+    crawler.engine = AsyncMock()
+
+    with mock.patch(
+        "scrapy_poet.page_input_providers.create_scrapy_backend"
+    ) as mock_factory:
+        provider = HttpClientProvider(crawler)
+        results = provider(set(), crawler)
+        assert isinstance(results[0], HttpClient)
+
+    results[0].request_downloader == mock_factory.return_value
+
+def test_meta_provider(settings):
+    crawler = get_crawler(Spider, settings)
+    provider = MetaProvider(crawler)
+    request = scrapy.http.Request("https://example.com")
+
+    results = provider(set(), request)
+
+    assert results[0] == {}
+
+    expected_data = {"key": "value"}
+    request.meta.update({"po_args": expected_data})
+    results = provider(set(), request)
+
+    assert results[0] == expected_data